Remove tile size limit
if previous modifications do not help, remove tile size limit. since these are optional tiles toggled by the user, we think we can have them be as large as necessary in order to accurately provide information rather than needing to pare them down to be as fast loading as possible.